// src/twelve-data/types.d.ts
// Generated by https://quicktype.io
import { IError } from '@app/types'

export interface TwelveDataPreviousCloseAttrs {
  symbol: string
  name: string
  exchange: string
  mic_code: string
  currency: string
  datetime: string
  timestamp: number
  open: string
  high: string
  low: string
  close: string
  volume: string
  previous_close: string
  change: string
  percent_change: string
  average_volume: string
  is_market_open: boolean
  fifty_two_week: TwelveDataFiftyTwoWeekAttrs
}

export interface TwelveDataFiftyTwoWeekAttrs {
  low: string
  high: string
  low_change: string
  high_change: string
  low_change_percent: string
  high_change_percent: string
  range: string
}

export interface Dictionary {
  [Key: string]: TwelveDataPreviousClose
}

export interface IApiError extends IError {
  status: 'error'
}